+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - New #50: Add `deleteStickerSet` method.
 - New #51: Add `getStickerSet` method and `StickerSet` type.
 - New #52: Add `sendSticker` method.
+- New #54: Add `getCustomEmojiStickers` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.